FREESCAPE: Move digital sound code into separate classes
