+++ engines/freescape/freescape.cpp
@@ -342,13 +342,6 @@ FreescapeEngine::~FreescapeEngine() {
 		delete it;
 	}
 
-	for (auto &it : _soundsFx) {
-		if (it._value) {
-			free(it._value->data);
-			free(it._value);
-		}
-	}
-
 	if (_savedScreen) {
 		_savedScreen->free();
 		delete _savedScreen;
