+++ engines/freescape/games/dark/atari.cpp
@@ -69,7 +69,7 @@ void DarkEngine::loadAssetsAtariFullGame() {
 	loadMessagesVariableSize(stream, 0x3f6f, 66);
 	loadPalettes(stream, 0x204d6);
 	loadGlobalObjects(stream, 0x32f6, 24);
-	loadSoundsFx(stream, 0x266e8, 11);
+	_sound = loadSoundsFx(stream, 0x266e8, 11);
 
 	for (auto &area : _areaMap) {
 		// Center and pad each area name so we do not have to do it at each frame
