+++ engines/freescape/games/eclipse/eclipse.h
@@ -109,7 +109,6 @@ public:
 	Common::String getScoreString(int score);
 	void drawScoreString(int score, int x, int y, uint32 front, uint32 back, Graphics::Surface *surface);
 
-	soundFx *load1bPCM(Common::SeekableReadStream *file, int offset);
 	void loadHeartFramesCPC(Common::SeekableReadStream *file, int restOffset, int beatOffset);
 	void loadHeartFramesZX(Common::SeekableReadStream *file, int restOffset, int beatOffset);
 	void loadHeartFramesDOS(Common::SeekableReadStream *file, int restOffset, int beatOffset);
@@ -181,11 +180,13 @@ public:
 	bool triggerWinCondition() override;
 	bool checkIfGameEnded() override;
 	void endGame() override;
-	void loadSoundsFx(Common::SeekableReadStream *file, int offset, int number) override;
 	void playSoundFx(int index, bool sync) override;
 
 	Common::Error saveGameStreamExtended(Common::WriteStream *stream, bool isAutosave = false) override;
 	Common::Error loadGameStreamExtended(Common::SeekableReadStream *stream) override;
+
+private:
+	Sound *_soundFx;
 };
 
 }
